+++ zhalilov.rail/P1/main.cpp
@@ -1,43 +1,11 @@
 #include <iostream>
-
-class Counter
-{
-  size_t count;
-  size_t maxSeq;
-public:
-  Counter():
-    count(1),
-    maxSeq(0)
-  {}
-
-  void operator()(int prev, int curr)
-  {
-    if(curr == prev)
-    {
-      count++;
-    }
-    else
-    {
-      count = 1;
-    }
-
-    if(count > maxSeq)
-    {
-      maxSeq = count;
-    }
-  }
-
-  size_t operator()() const
-  {
-    return maxSeq;
-  }
-};
+#include "counter.hpp"
 
 int main()
 {
   int prev = 0;
   int curr = 0;
-
+  using namespace zhalilov;
   Counter counter;
   while(std::cin >> curr && curr)
   {
